added core order matcher

+++ src/main/java/com/coindcx/api/persistance/POJO/request/AddOrderRequest.java
@@ -3,13 +3,14 @@ package com.coindcx.api.persistance.POJO.request;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.sql.Time;
 
 @Getter
 @Setter
 public class AddOrderRequest {
-    private Long size;
-    private Long price;
+    private BigDecimal size;
+    private BigDecimal price;
     private Time time;
     private String side;
 }
